+++ src/silicone/time_projectors/__init__.py
@@ -1,0 +1,9 @@
+"""
+Time projectors
+
+The classes in this module are used to infer values for a scenario at later times given
+the trends before that time.
+
+"""
+
+from .extend_latest_time_quantile import ExtendLatestTimeQuantile  # noqa: F401