+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Work in progress
 ----------------
 Added
 ~~~~~
-- (`#126 <https://github.com/GranthamImperial/silicone/pull/126>`_) Added a cruncher (Extend latest time quantile) that extends a pathway to cover later times, assuming it remains at the same quantile. (Like EQW but for a single variable)
+- (`#126 <https://github.com/GranthamImperial/silicone/pull/126>`_) Added the first time projector (Extend latest time quantile) that extends a pathway to cover later times, assuming it remains at the same quantile. 
 
 [v1.0.3]
 --------